initial version 0.1 of "cr_orapcs.sh"
Initial version 0.1 of the Azure CLI bash script "cr_orapcs.sh", which fully automates the creation of a Pacemaker/Corosync (PCS) cluster for an Oracle v12.2 Standard Edition database on a primary host and a failover host within Azure.  Prerequisites are an Azure subscription (of course), an Azure resource group, and an Azure CLI version 2.50 or higher.  Also, at the present time, Azure shared disk is available only in West Central US region.